+++ QoL/GlobeTweaks/GlobeTweaks.cs
@@ -1,13 +1,4 @@
-﻿using Base.Core;
-using PhoenixPoint.Geoscape.Entities;
-using PhoenixPoint.Geoscape.Entities.Abilities;
-using PhoenixPoint.Geoscape.Levels;
-using PhoenixPoint.Geoscape.View.ViewControllers;
-using PhoenixPoint.Geoscape.View.ViewModules;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
+﻿using System;
 
 namespace Sheepy.PhoenixPt.GlobeTweaks {
 
@@ -37,6 +28,7 @@ namespace Sheepy.PhoenixPt.GlobeTweaks {
                Auto_Unpause_Multiple = true;
             ZyMod.Api( "config save", this );
          }
+         if ( Haven_Icons == null ) Haven_Icons = new HavenIconConfig();
       }
    }
 
@@ -60,79 +52,10 @@ namespace Sheepy.PhoenixPt.GlobeTweaks {
 
       public void GeoscapeMod ( Func< string, object, object > api = null ) {
          SetApi( api, out Config ).Upgrade();
-         if ( Config.Show_Airplane_Action_Time )
-            TryPatch( typeof( UIModuleSiteContextualMenu ), "SetMenuItems", postfix: nameof( AfterSetMenuItems_CalcTime ) );
-         if ( Config.Same_Site_Scan_Cooldown_Min != 0 )
-            TryPatch( typeof( UIModuleSiteContextualMenu ), "SetMenuItems", postfix: nameof( AfterSetMenuItems_DisableDupScan ) );
+         new TimeModule().DoPatches();
          new PauseModule().DoPatches();
          new GlyphModule().DoPatches();
       }
 
-      private static void AfterSetMenuItems_CalcTime ( GeoSite site, List<SiteContextualMenuItem> ____menuItems ) {
-         foreach ( var menu in ____menuItems ) try {
-            var vehicle = menu.Ability?.GeoActor as GeoVehicle;
-
-            if ( menu.Ability is MoveVehicleAbility move && move.GeoActor is GeoVehicle flier )
-               menu.ItemText.text += HoursToText( FlightHours( site, flier ) );
-
-            else if ( menu.Ability is ExploreSiteAbility explore ) {
-               var hours = GetVehicleTimeLeft( vehicle, (float) site.ExplorationTime.TimeSpan.TotalHours );
-               menu.ItemText.text += HoursToText( hours );
-
-            } else if ( menu.Ability is ScanAbility scan ) {
-               var scanner = GetScanner( site, scan, out float hours, true );
-               if ( scanner != null ) menu.ItemText.text += HoursToText( hours );
-            } // LaunchMissionAbility, HavenTradeAbility, HavenDetailsAbility
-         } catch ( Exception ex ) { Error( ex ); }
-      }
-
-      private static void AfterSetMenuItems_DisableDupScan ( GeoSite site, List<SiteContextualMenuItem> ____menuItems ) { try {
-         foreach ( var menu in ____menuItems ) {
-            if ( ! ( menu.Ability is ScanAbility scan ) ) continue;
-            if ( ! menu.Button.IsInteractable() ) return;
-            var scanner = GetScanner( site, scan, out float hours, false );
-            if ( scanner == null ) return;
-            var duration = scanner.ScannerDef.ExpansionTimeHours - hours;
-            var cooldown = Config.Same_Site_Scan_Cooldown_Min / 60f;
-            if ( cooldown == 0 || ( cooldown > 0 && duration > cooldown ) ) return;
-            Verbo( "Preventing duplicate scan.  Last scan {0}hr, threshold {1}hr.", duration, cooldown );
-            menu.Button.SetInteractable( false );
-            break;
-         }
-      } catch ( Exception ex ) { Error( ex ); } }
-
-      private static GeoScanner GetScanner ( GeoSite site, GeoAbility ability, out float remainingHours, bool GetFirst ) {
-         remainingHours = 0;
-         var scanners = ( ability.GeoActor as GeoVehicle )?.Owner?.Scanners?.Where( e => e.Location == site );
-         if ( scanners == null || ! scanners.Any() ) return null;
-         var scanner = GetFirst ? scanners.First() : scanners.Last();
-         remainingHours = (float) ( ( scanner.SerializationData as GeoScanner.GeoScannerInstanceData ).ExpansionEndDate - scanner.Timing.Now ).TimeSpan.TotalHours;
-         return scanner;
-      }
-
-      private static float GetVehicleTimeLeft ( GeoVehicle vehicle, float def ) { try {
-         if ( vehicle == null ) return def;
-         var updateable = typeof( GeoVehicle ).GetField( "_explorationUpdateable", BindingFlags.NonPublic | BindingFlags.Instance ).GetValue( vehicle );
-         if ( updateable == null ) return def;
-         NextUpdate endTime = (NextUpdate) updateable.GetType().GetProperty( "NextUpdate" ).GetValue( updateable );
-         return (float) -( vehicle.Timing.Now - endTime.NextTime ).TimeSpan.TotalHours;
-      } catch ( Exception ex ) { Warn( ex ); return def; } }
-
-      private static float FlightHours ( GeoSite site, GeoVehicle vehicle ) {
-         var pos = vehicle.CurrentSite?.WorldPosition ?? vehicle.WorldPosition;
-         var path = vehicle.Navigation.FindPath( pos, site.WorldPosition, out bool hasPath );
-         if ( ! hasPath || path.Count < 2 ) return -1;
-         float distance = 0;
-         for ( int i = 0, len = path.Count - 1 ; i < len ; )
-            distance += GeoMap.Distance( path[i].Pos.WorldPosition, path[++i].Pos.WorldPosition ).Value;
-         return distance / vehicle.Stats.Speed.Value;
-      }
-
-      private static string HoursToText ( float hours ) {
-         if ( hours < 0 ) return "-" + HoursToText( -hours );
-         int intHour = (int) hours;
-         int d = Math.DivRem( intHour, 24, out int h ), m = (int) Math.Ceiling( ( hours - intHour ) * 60 );
-         return string.Format( d > 0 ? Config.Days_Format : Config.Hours_Format, d, h, m, hours, hours * 60 );
-      }
    }
 }